Javítások, pár extra dolog és most már postmannal mnnek a kérések. Viszont 500 hiba történik mindig

+++ Backend/BidVerse_backend/src/main/kotlin/org/example/bidverse_backend/repositories/CategoryRepository.kt
@@ -3,7 +3,7 @@ package org.example.bidverse_backend.repositories
 import org.example.bidverse_backend.entities.Category
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface CategoryRepository : JpaRepository<Category, Long>{
+interface CategoryRepository : JpaRepository<Category, Int>{
     fun findByIdIn(categoryIds: List<Int>): List<Category>
 
     fun findByCategoryName(categoryName: String): Category?
